notes_app/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
distlib==0.3.3
Django==3.2.7
django-rest-framework==0.1.0
djangorestframework==3.12.4
djongo==1.3.6
filelock==3.3.2
platformdirs==2.4.0
pymongo==3.12.0
pytz==2021.1
six==1.16.0
sqlparse==0.2.4
virtualenv==20.10.0
